divide la pagina en componentes
