+++ components/contacto.html
@@ -1,0 +1,1 @@
+<h1>Contacto</h1>
