+++ components/inicio.html
@@ -1,0 +1,1 @@
+<h1>Inicio</h1>
