+++ components/nosotros.html
@@ -1,0 +1,1 @@
+<h1>Nosotros</h1>
