+++ components/proyectos.html
@@ -1,0 +1,1 @@
+<h1>Proyectos</h1>
